// src/helpers/types.d.ts
export type GQLResponseType = {
    cursor: string | null;
    node: {
        id: string;
        tags: KeyValueType[];
        data: {
            size: string;
            type: string;
        };
    };
};
// ...
export interface PoolType {
    id: string;
    state: PoolStateType;
}
export interface PoolStateType {
    title: string;
    image: string;
    briefDescription: string;
    description: string;
    link: string;
    owner: string;
    ownerInfo: string;
    timestamp: string;
    contributors: {
        [key: string]: string;
    };
    tokens: {
        [key: string]: string;
    };
    totalContributions: string;
    totalSupply: string;
    balance: string;
}
// ...
export type ContributionResultType = {
    status: boolean;
    message: string | null;
};
// ...
export type KeyValueType = {
    [key: string]: string | any;
};
// ...
export type ContributionType = {
    timestamp: string;
    qty: string;
};